:bug: :sparkles: :boom: :art: Levels are now generated in node
Editing src/levels.h manually was annoying, so I decided that it was
time to make the level creation and edition easier. I have chosen for
the sake of simplicity to create a level_X.js file, which will then be
ran to create a level structure, that gets transpiled in a .h file,
following the structure of the previous src/levels.h.

Amongst the other changes included:
* fixed types in src/graphics/color.h
* draw GHOST_LINE like a TRANSPARENT_LINE
* created a few other, test levels (as it is now way easier)